python/langchain-ampersend/pyproject.toml
[project]
name = "langchain-ampersend"
version = "0.0.1"
description = "LangChain integration for Ampersend x402 payments"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-core>=0.3.0",
    "ampersend-sdk",
    "httpx>=0.27.0",
]

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "langchain-tests==1.1.2",  # Pin to avoid breaking CI
]
